fix: 분석 로직 수정

+++ backend/src/main/java/com/buildsync/dto/analysis/MonthlyMaterialCostResponse.java
@@ -8,6 +8,6 @@ import lombok.Getter;
 public class MonthlyMaterialCostResponse {
 
 	private String month; // 월
-	private Integer totalQuantity; // 발주 총 금액
-	private Integer totalMaterialCost; // 총 자재 구매 비용
+	private Long totalQuantity; // 발주 총 금액
+	private Long totalMaterialCost; // 총 자재 구매 비용
 }
